Fixes default GUI viewer pose for G1 locomanip task (#187)
## Summary
The default camera view of the G1 locomanip task was placing camera
outside the walls of the scene, leading users to only see a grey wall
when the env is loaded up.

This PR updates the default viewer position so that the inside of the
galileo lab is seen by default.